MoviePeople moved to common folder
